3/1/18
BuySuppliesView - Corrected handling when an invalid menu option is entered so that it displays and error message instead of asking for an item amount to purchase and then showing the error message.

Corrected if-else-if ladder to get index number for each inventory item.

Added try and catch for error when a letter is entered instead of a number at the buy amount prompt.

Used inventoryType.inventoryName from enum instead of creating a variable with the information for the success message.

InventoryItem – Changed inventoryType to  inventoryName. Correct some of the names in the enum data.