3/9/18
Added "Please Choose an Option" to overloaded constructor for GeneralStoreView, HelpMenuView, and MainMenuView.

GeneralStoreView we had @Override for the getInput and we didn't need it, so I deleted it.

Deleted commented code in GeneralStoreView and BuySuppliesView.

Modified MainMenuView to add option for RiverCrossingView to menu.
Modified MapControl to add moveTeam stub method.
Modified GameControl to add determineWin stub method.

Added CrossRiverView, DropSuppliesView, RiverCrossingView.